README: Replace local contrib docs with a CONTRIBUTING.md link
CONTRIBUTING.md (which we got from project-template in be21fe9, *:
Pull in project-template 61d73a3, 2018-04-04, #3) covers this same
space.  Dropping the contrib docs from the README (and instead linking
to CONTRIBUTING.md) keeps us DRY (vs. having the content in each
location), gets us the GitHub contributing UI improvements [1]
(vs. having the content only in the README), keeps the information
discoverable (vs. having the content only in CONTRIBUTING and not
linking from the README), and gives us more space in the README to
talk about the project itself (contributing docs are only useful for a
subset of README readers).

I've copy/pasted the CONTRIBUTING.md content for issues and the
mailing list into the README at Stephen's request [2], turning
"Issues" into a link in the process and making some other minor
copy-edits.  I've left it as a non-link in CONTRIBUTING.md to stay
closer to the upstream project-template, but wouldn't have a problem
making it a link there to if the maintainers preferred that.  I've
also added a security reference based on my in-flight [3], also at
Stephen's request [4].  Personally I don't have a problem with
requiring readers to click through to CONTRIBUTING to get this
information (as long as the security information stays close to the
information for other channels [3]), but I'm not a maintainer.

[1]: https://help.github.com/articles/setting-guidelines-for-repository-contributors/
[2]: https://github.com/opencontainers/distribution-spec/pull/9#discussion_r180902283
[3]: https://github.com/opencontainers/project-template/pull/57
[4]: https://github.com/opencontainers/distribution-spec/pull/9#discussion_r180915017

Signed-off-by: W. Trevor King <wking@tremily.us>