Correct grammar for digest algorithm portion
The current grammar specifies that the characters [A-Fa-f0-9_+.-] are
valid in the "algorithm" portion of a digest. This contradicts the
specification stating that "compliant implementations SHOULD use
sha256", as "sha256" does not match the above rule.

This commit fixes this issue by additionally allowing characters g-zG-Z
in the "algorithm" portion.

Signed-off-by: Fabian Freyer <fabian.freyer@physik.tu-berlin.de>