Update blob patch status code to correct code
The PATCH blob upload endpoint returns a 202 Accepted when a chunk
is accepted rather than a 204. The 204 is returned by the PUT
endpoint after completion of an upload is successful. The status
code in the detail section must match the protocol as described
in the description for upload patching. It is stated at the beginning
of the detail section that when the detail does not match, the
description above it is correct.

Signed-off-by: Derek McGowan <derek@mcgstyle.net>