Introduce OCI Conformance Test Suite
Added new conformance directory in the project root, with a number of
test files written in Go. Tests can be compiled by running `go test -c`
in the conformance directory and executing the created conformance.test
file.

In order for the tests to run, registry providers will need to set up
certain environment variables with the root url, the namespace of a
repository, and authentication information. Additionally, the OCI_DEBUG
variable can be set to "true" for more detailed output.

The tests create two report files: report.html and junit.xml. The html
report is expandable if more detailed information is needed on failures.

Related to #24

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>