Proof of concept
I set up a system through which we can configure options and sub-options
from the command line. Several constants and maps were defined to make
the code readable. The main configuration options (OCI_TEST_PUSH, etc)
require several sub-options which are validated before the tests run. If
they are not all present, we exit with an error and a report of what was
supplied vs what wasn't.

Tests can now be skipped like so, in the `Specify` block:
```
SkipIfNotEnabled(push)
```

We're still in need of tests to set up combinations of configurations
that are guaranteed to work, and we still need to actually split up the
test flow.

Signed-off-by: Josh Dolitsky <393494+jdolitsky@users.noreply.github.com>