spec: add missing TOOMANYREQUESTS (429) error-code
The `TOOMANYREQUESTS` error code was added in docker/distribution#1693,
and added to the documentation in docker/distribution#1833.

While it's documented in various examples, it never made it to the list
of allowed error-codes.

This patch is fixing that omission and adding it to the list of valid
error-codes that can be returned by registries

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>