Remove detail.md
Although detail.md contained some useful information, the majority
of the document contained information that wasn't relevant to a
generic OCI registry.

It contained extraneous information, like information about authentication
that many registries may not use. In addition, the defined behaviours
were not conformant to the specification.

The spec still has a permalink to a github version of the detail.md.

Signed-off-by: Sargun Dhillon <sargun@sargun.me>