Allow Content-Length to be omitted when pushing on patch requests
Technically due to the way the specification is written today the moby/docker
cli push is nonconformant. Given that is one of the most  common ways users
are pushing today, the specification should recognize that it is allowed.

Further, requiring `Content-Length` to be set means that no client that
is conforming can use http chunked encoding, which limits the use of
CDNs like CloudFlare and other reverse proxies. See further discussion
at: https://github.com/containerd/containerd/pull/7459

Signed-off-by: Ian Zink <ian@replicated.com>