conformance: check response status before checking location
I noticed that when these tests are running against a server
that returns the wrong status code (an error, for example),
the failure that we see doesn't mention the status code, but
only that the returned location is empty.

This change checks the status code first, as if that's wrong, the
location is very unlikely to be relevant.

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>