conformance: support OCI_REPORT_DIR
This makes it possible to run the conformance tests in a read-only directory
without failing by specifying an output directory for the report files
instead of hard-coding it to the current directory (which may not be
writable, and _won't_ be writable if the tests are being run with
`go test github.com/opencontainers/distribution-spec/conformance`
rather than in a git checkout of the conformance module.

We recognize the value `none` for disabling report generation entirely.

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>