Restore pagination information
Notably "breaking" changes here are that the registry may impose a limit
on the number of tags it returns.

This was restored from b4e98332eccaf51d1974c9136caf67317bd99104, where
it applied to Catalog, which was removed, but given that many registries
have `adopted this behavior already, we should document it correctly.

This also restores language that clients SHOULD always use the Link
header when present.

This was restored from c90b0f145ac6bc09d2636ee214486ac333edc284.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>