Conformance redesign
This reworks the conformance tests to verify the various APIs with different sets of data.
The configuration allows various APIs and types of data to be disabled.
The configuration will default to the tests appropriate for a specific version of the spec.
However, new types of data may still be pushed when testing older versions of the spec.

The tests are also designed to be run with `go build` rather than `go test`.
This will allow the conformance tests themselves to be tested in the future.

Attempts to run the tests with `go test` and the previous configuration variables should still work, but support for this is deprecated and warning will be included in the output to update.

The results now include a new results.yaml including the configuration and the result of each API and type of data tested.
Processing this on the oci-conformance repo requires a separate PR there.

This redesign was necessary to allow the conformance tests to scale with the different types of data we are seeing.
This includes data with a subject, artifacts, and new digest algorithms.

Signed-off-by: Brandon Mitchell <git@bmitch.net>