Refactor blob creation and deletion
Define functions to create and delete blobs. This way multiple issues
are addressed:
- Code readability improves a lot.
- Response is strictly validated.
- Keep track of created blobs to cleanup them all during teardown.

Previously list of blobs to be cleanup was manually maintained and was
error prone.
Blobs are deleted in the reverse creation order to avoid dependency
issues.

This is a pure refactoring, no external changes can be observed.

Signed-off-by: Silvano Cirujano Cuesta <silvano.cirujano-cuesta@siemens.com>